+++ lib/src/devices/a3936/packets.rs
@@ -1,3 +1,5 @@
+mod set_custom_button_model_packet;
 mod state_update_packet;
 
+pub use set_custom_button_model_packet::*;
 pub use state_update_packet::*;